Fix rewriter when the sync method is in the same extensions class
Pull request #108 fixed EF Core async to sync generation. EF Core async methods are found in the `EntityFrameworkQueryableExtensions` class, so a mapping from EntityFrameworkQueryableExtensions to System.Linq.Queryable was added.

Unfortunately, this was not enough. The `EntityFrameworkQueryableExtensions` class also contains extension methods that have both an async and a sync method. For example: `ExecuteDeleteAsync` and `ExecuteDelete`.

This commit fixes the translation for those methods where both the async and sync methods are in the same extension class.